[AOS] refactor: Data Layer 폴더 정리
