+++ AOS/app/src/main/java/com/hyundai/myexperience/data/remote/data_source/ReservationQueueDataSource.kt
@@ -1,4 +1,4 @@
-package com.hyundai.myexperience.data.remote
+package com.hyundai.myexperience.data.remote.data_source
 
 import com.hyundai.myexperience.data.remote.client.ReservationClient
 import kotlinx.coroutines.flow.Flow
