+++ AOS/app/src/main/java/com/hyundai/myexperience/data/repository/NoticeRepository.kt
@@ -1,10 +1,10 @@
-package com.hyundai.myexperience.data
+package com.hyundai.myexperience.data.repository
 
 import com.hyundai.myexperience.data.entity.notice.NoticeDetailItem
 import com.hyundai.myexperience.data.entity.notice.NoticesItem
 import com.hyundai.myexperience.data.mapper.mapToNoticeList
 import com.hyundai.myexperience.data.mapper.mapToNoticesItem
-import com.hyundai.myexperience.data.remote.NoticeRemoteDataSource
+import com.hyundai.myexperience.data.remote.data_source.NoticeRemoteDataSource
 import javax.inject.Inject
 
 class NoticeRepository @Inject constructor(
