+++ AOS/app/src/main/java/com/hyundai/myexperience/ui/joined_program/JoinedProgramViewModel.kt
@@ -4,11 +4,8 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.hyundai.myexperience.data.JoinedProgramRepository
+import com.hyundai.myexperience.data.repository.JoinedProgramRepository
 import com.hyundai.myexperience.data.entity.my_page.JoinedProgramItem
-import com.hyundai.myexperience.utils.getCompanyName
-import com.hyundai.myexperience.utils.getLevel
-import com.hyundai.myexperience.utils.getProgramName
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.launch
 import javax.inject.Inject
