+++ AOS/app/src/main/java/com/hyundai/myexperience/ui/notice/NoticeListViewModel.kt
@@ -4,7 +4,7 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.hyundai.myexperience.data.NoticeRepository
+import com.hyundai.myexperience.data.repository.NoticeRepository
 import com.hyundai.myexperience.data.entity.notice.NoticesItem
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.launch
