+++ AOS/app/src/main/java/com/hyundai/myexperience/ui/reservation_entrance/ReservationEntranceViewModel.kt
@@ -4,8 +4,8 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.hyundai.myexperience.data.ReservationRepository
-import com.hyundai.myexperience.data.UserRepository
+import com.hyundai.myexperience.data.repository.ReservationRepository
+import com.hyundai.myexperience.data.repository.UserRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.launch
 import javax.inject.Inject
