+++ AOS/app/src/main/java/com/hyundai/myexperience/ui/signin/SignInViewModel.kt
@@ -2,9 +2,8 @@ package com.hyundai.myexperience.ui.signin
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.hyundai.myexperience.data.UserRepository
+import com.hyundai.myexperience.data.repository.UserRepository
 import com.hyundai.myexperience.data.remote.ServerConnection
-import com.hyundai.myexperience.utils.showToast
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.launch
 import javax.inject.Inject
