+++ AOS/app/src/main/java/com/hyundai/myexperience/ui/signup/SignUpViewModel.kt
@@ -2,7 +2,7 @@ package com.hyundai.myexperience.ui.signup
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.hyundai.myexperience.data.UserRepository
+import com.hyundai.myexperience.data.repository.UserRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.launch
 import javax.inject.Inject
